bindings: split up python bindings
This commit contains only code movement.
